Allow to configure the imageRenderer as valuebinding (in addition to the ImageRenderer class name as string). Encode the url to get in the session id and conversationContext (eventually) into the url.
This allows the developer to use any managed bean to provide the image data.

<s:graphicImageDynamic width=".." height=".." value="#{bean.imageData}" />

imageData is a method with the following signature:

public ImageRenderer getImageData()

This works with any session, application, conversation scoped bean.
In fact it works with request scoped beans too, but then the pro to have an already initialized bean is lost.

git-svn-id: http://svn.apache.org/repos/asf/myfaces/tomahawk/trunk@544554 13f79535-47bb-0310-9956-ffa450edef68