Add a way to check if dojo is going to be used within the rendered page. The check will return true after the main dojo initialization has been added to the dojo scripts.
Make use of dojo (if it is already in use) in SubmitOnEventRenderer to check if dojo.addOnLoad can be used insted of the setTimeout way which makes problems with the dojo modal dialog widget (timeout might be too short before the widget has been created).
Avoid recursive adding of onHide method to the dialog component (makes it work with ppr).
Notice: SubmitOnEvent uses dojo only if there is already a component requireing it. This avoids loading the heavy-weight library when alternatives exist.

git-svn-id: http://svn.apache.org/repos/asf/myfaces/tomahawk/trunk@572241 13f79535-47bb-0310-9956-ffa450edef68